Set abbreviations keypad elements
