+++ app/modules/transcribe/overlay/keypad-slide.html
@@ -1,0 +1,5 @@
+<div class="keypad-slide">
+    <div ng-class="visible ? 'sidebar -visible' : 'sidebar'">
+        Keypad
+    </div>
+</div>
