+++ app/modules/transcribe/overlay/overlay.html
@@ -2,5 +2,6 @@
     <div loading-spinner></div>
     <div context-menu></div>
     <div transcribe-dialog></div>
+    <div keypad-slide></div>
     <div out-of-data></div>
 </div>
