+++ app/styles/components/keypad-slide.styl
@@ -1,0 +1,14 @@
+.keypad-slide
+
+    > .sidebar
+        transition: right .5s
+        width: 100px
+        background-color: blue
+        position: absolute
+        top: 0px
+        bottom: 0px
+        right: -100px
+
+
+    > .sidebar.-visible
+        right: 0px
