+++ app/styles/components/transcribe-dialog.styl
@@ -54,6 +54,9 @@
     > .body > .buttons .button.-save
         @extend .btn-success
 
+    > .body > .buttons .button.-keypad
+        @extend .btn-primary
+
     > .body > .buttons .button + .button
         margin-left: -1px
